# Improve performance of UUID parsing