--- a/specs/Qowaiv.Benchmarks/Program.cs
+++ b/specs/Qowaiv.Benchmarks/Program.cs
@@ -7,6 +7,8 @@ public static class Program
     public static void Main()
     {
         BenchmarkRunner.Run<UuidBenchmark.Parse>();
-        //BenchmarkRunner.Run<IbanBenchmark.ParseFormatted>();
+        BenchmarkRunner.Run<UuidBenchmark.Version>();
+        BenchmarkRunner.Run<IbanBenchmark.ParseUnformatted>();
+        BenchmarkRunner.Run<IbanBenchmark.ParseFormatted>();
     }
 }