--- a/specs/Qowaiv.Benchmarks/Program.cs
+++ b/specs/Qowaiv.Benchmarks/Program.cs
@@ -6,9 +6,9 @@ public static class Program
 {
     public static void Main()
     {
-        BenchmarkRunner.Run<UuidBenchmark.Parse>();
-        BenchmarkRunner.Run<UuidBenchmark.Version>();
-        BenchmarkRunner.Run<IbanBenchmark.ParseUnformatted>();
-        BenchmarkRunner.Run<IbanBenchmark.ParseFormatted>();
+        BenchmarkRunner.Run<UuidBenchmark.StringOutput>();
+        //BenchmarkRunner.Run<UuidBenchmark.Version>();
+        //BenchmarkRunner.Run<IbanBenchmark.ParseUnformatted>();
+        //BenchmarkRunner.Run<IbanBenchmark.ParseFormatted>();
     }
 }