--- a/specs/Qowaiv.Benchmarks/Qowaiv.Benchmarks.csproj
+++ b/specs/Qowaiv.Benchmarks/Qowaiv.Benchmarks.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.11" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.*" />
     <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
   </ItemGroup>