… specs/Qowaiv.Benchmarks/README.md
… src/Qowaiv.Data.SqlClient/Generated/Timestamp.generated.cs
… src/Qowaiv/Generated/Chemistry/CasRegistryNumber.generated.cs
… src/Qowaiv/Generated/Date.generated.cs
… src/Qowaiv/Generated/DateSpan.generated.cs
… src/Qowaiv/Generated/EmailAddress.generated.cs
… src/Qowaiv/Generated/Financial/Amount.generated.cs
… src/Qowaiv/Generated/Financial/BusinessIdentifierCode.generated.cs
… src/Qowaiv/Generated/Financial/Currency.generated.cs
… src/Qowaiv/Generated/Financial/InternationalBankAccountNumber.generated.cs
… src/Qowaiv/Generated/Financial/Money.generated.cs
… src/Qowaiv/Generated/Globalization/Country.generated.cs
… src/Qowaiv/Generated/HouseNumber.generated.cs
… src/Qowaiv/Generated/IO/StreamSize.generated.cs
… src/Qowaiv/Generated/LocalDateTime.generated.cs
… src/Qowaiv/Generated/Mathematics/Fraction.generated.cs
… src/Qowaiv/Generated/Month.generated.cs
… src/Qowaiv/Generated/MonthSpan.generated.cs
… src/Qowaiv/Generated/Percentage.generated.cs
… src/Qowaiv/Generated/PostalCode.generated.cs
… src/Qowaiv/Generated/Sex.generated.cs
… src/Qowaiv/Generated/Statistics/Elo.generated.cs
… src/Qowaiv/Generated/Sustainability/EnergyLabel.generated.cs
… src/Qowaiv/Generated/Uuid.generated.cs
… src/Qowaiv/Generated/Web/InternetMediaType.generated.cs
… src/Qowaiv/Generated/WeekDate.generated.cs
… src/Qowaiv/Generated/Year.generated.cs
… src/Qowaiv/Generated/YesNo.generated.cs
… src/Qowaiv/Identifiers/GuidLayout.cs
… src/Qowaiv/Identifiers/GuidParser.cs